1. added all four functions 2. rearranged code to reuse tf read

+++ qubole/entity/StableSpotInstanceSettings.go
@@ -5,7 +5,7 @@ import (
 )
 
 type StableSpotInstanceSettings struct {
-	Maximum_bid_price_percentage     int	`json:"maximum_bid_price_percentage,omitempty"`
+	Maximum_bid_price_percentage     float32	`json:"maximum_bid_price_percentage,omitempty"`
 	Timeout_for_request              int	`json:"timeout_for_request,omitempty"`
 }
